move to spark operation
